Create components.py with first draft of Cell class.
The components.py file is intended to hold the "agents" of the
game (Cells), the "world" where they live in (Board) and the way
they interact according to the rules of the Game.

The Cell class describes the "agents", the basic block of the
Game that fill the Board. Cell class principal attribute is
.state, a boolean variable taken as input as "ON" or "OFF".

At the moment, Cell instances can just be created, have a
.state value assigned, changed or printed. Next implementation
after testing for bugs and errors is adding a history track.
Documenting the file is top priority.