Simplify Cell class and create Board class basic working.
The class Cell has been simplied to its basics. The `.status`
attribute is set as False by default and the only method,
`.turn_on()` turns the state to True. The proposals from the
previous commit have been posted as Issue #1.

Te class Board takes as argument the intger `n` when instantiated
or is set to n=5  by default. This argument is used to create
an n x n grid of Cell objects turned off. The methods work to
reproduce an ascii prototype of the game.

Documentation is still required ASAP